before: replace the bitvec build buffer with the crate-owned BitsBuf
The build side's one remaining 32-bit boundary was the bit vector behind
the emitters: its length encoding caps at usize::MAX >> 3 bits, so an
emitting operation whose output crossed 2^29 bits (a covered join of a
64 MiB operand, a two-operand join concatenating past it) trapped at the
freeze seam on wasm32 while every walk and door was already exact.

BitsBuf is the one build buffer now: bytes beside a u64 live length,
under two representation invariants every mutation maintains — the byte
vector holds exactly the live bits' bytes, and dead bits in the final
partial byte are zero (truncation masks the new final byte in O(1)).
The byte image is therefore a function of the bit content alone: Eq is
one memcmp, sealing appends only the marker, and the freeze seam adopts
the allocation whole. PackedBuilder keeps its staged-register internals
and word-parallel moves but runs its positions at u64 and hands its
finished bytes straight to BitsBuf — no length-encoding conversion binds
the hand-off on any target. The storable bound (usize-denominated bit
positions in the frozen form) is asserted loudly at the freeze door,
matching the decode door's existing contract. The gamma word path emits
through BitsBuf::push_bits; bitvec leaves the production dependency
graph (it stays a dev-dependency as emit_probe's external baseline).

The build-history proptest family pins the invariants as a family:
arbitrary interleavings of pushes, word appends, view copies, and
truncations (byte-aligned, mid-byte, to-empty, multi-byte sheds) leave
the buffer byte-identical to a clean rebuild of the surviving content at
every intermediate state, freeze to the identical canonical spelling,
and keep Eq and the canonical hash agreeing with bit-level equality;
a second leg pins spelling injectivity across distinct contents. A
transient no-zeroing mutant of truncate fails both legs (adequacy
verified before landing; the mutant is not committed).

wasm32-pins: the three output-side trap pins flip green in this same
commit, per the pins file's own discipline — the covered joins at and
past 2^29 bits now assert byte-identical emission, and the two-operand
emit at the coordinate asserts its exact 536870905-bit output. The
boundary prose across the pins re-denominates to the straddle
coordinate (the bit count where a 32-bit length encoding would bind),
which every surface is now exact across. The rank pins stay as found
for the later phase that owns them.

All 36 wasm32 pins pass; the resource-envelope suite passes with every
pinned ceiling and floor untouched.

+++ crates/before/src/meter/tier2/tests.rs
@@ -385,7 +385,7 @@ fn skyline_join(a: &Version, b: &Version) -> Version {
     let decoded =
         skyline::decode(crate::codec::built_view(&out)).expect("an emitted join is canonical");
     assert_eq!(
-        out.len() as u64,
+        out.len(),
         tier2_size(crate::codec::built_view(&packed_bits_of(
             &to_oracle_version(&decoded)
         )))
@@ -404,7 +404,7 @@ fn skyline_meet(a: &Version, b: &Version) -> Version {
     let decoded =
         skyline::decode(crate::codec::built_view(&out)).expect("an emitted meet is canonical");
     assert_eq!(
-        out.len() as u64,
+        out.len(),
         tier2_size(crate::codec::built_view(&packed_bits_of(
             &to_oracle_version(&decoded)
         )))
